airflow/requirements.txt
kafka-python==2.0.2
pymongo==4.6.0
docker==6.1.3